+++ passnail.core/conf/generator.properties
@@ -1,4 +1,4 @@
-#Tue Feb 23 01:14:46 CET 2021
+#Tue Feb 23 21:43:50 CET 2021
 gen.upper.case.nr=1
 gen.lower.case.nr=5
 gen.passw.length.nr=9
