+++ passnail.core/pom.xml
@@ -52,11 +52,6 @@
         </dependency>
         <dependency>
             <groupId>com.passnail</groupId>
-            <artifactId>passnail.data.model</artifactId>
-            <version>${project.parent.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.passnail</groupId>
             <artifactId>passnail.data.service</artifactId>
             <version>${project.version}</version>
         </dependency>
@@ -82,6 +77,7 @@
                 <plugin>
                     <groupId>org.apache.maven.plugins</groupId>
                     <artifactId>maven-compiler-plugin</artifactId>
+                    <version>${version.compiler.plugin}</version>
                     <configuration>
                         <source>11</source>
                         <target>11</target>
@@ -105,7 +101,7 @@
                             <goal>repackage</goal>
                         </goals>
                         <configuration>
-                            <mainClass>com.passnail.server.core.app.ClientStarter</mainClass>
+                            <mainClass>com.passnail.core.main.ClientStarter</mainClass>
                         </configuration>
                     </execution>
                 </executions>
