+++ passnail.core/src/main/java/com/passnail/core/main/ClientStarter.java
@@ -1,0 +1,28 @@
+package com.passnail.core.main;
+
+import com.passnail.core.main.test.AttributesInitializer;
+import lombok.extern.log4j.Log4j2;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+
+/**
+ * A main class of the client's application.
+ */
+@Log4j2
+@SpringBootApplication
+@EnableJpaRepositories
+public class ClientStarter {
+
+    private static AttributesInitializer init;
+
+    public static void main(String[] args) {
+        init = new AttributesInitializer();
+        init.setAttributes(args);
+
+        SpringApplication.run(ClientStarter.class, args);
+
+    }
+
+}
