+++ passnail.core/src/main/java/com/passnail/core/throwable/IncorrectPropertiesException.java
@@ -1,0 +1,18 @@
+package com.passnail.core.throwable;
+
+/**
+ * An exception being thrown when the properties were not set correctly.
+ * <p>
+ * Created by: Pszemko at czwartek, 21.01.2021 21:57
+ * Project: passnail-client
+ */
+public class IncorrectPropertiesException extends IllegalArgumentException {
+    public IncorrectPropertiesException(final String s) {
+        super(s);
+    }
+
+    public IncorrectPropertiesException() {
+        super("Incorrect properties!");
+    }
+
+}
