--- passnail.core/src/main/java/com/passnail/server/core/app/ClientStarter.java
@@ -1,28 +1,0 @@
-package com.passnail.server.core.app;
-
-import com.passnail.server.core.app.config.AttributesInitializer;
-import lombok.extern.log4j.Log4j2;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-
-/**
- * A main class of the client's application.
- */
-@Log4j2
-@SpringBootApplication
-@EnableJpaRepositories
-public class ClientStarter {
-
-    private static AttributesInitializer init;
-
-    public static void main(String[] args) {
-        init = new AttributesInitializer();
-        init.setAttributes(args);
-
-        SpringApplication.run(ClientStarter.class, args);
-
-    }
-
-}
