+++ passnail.core/src/main/resources/application.properties
@@ -1,9 +1,6 @@
 spring.datasource.initialization-mode=always
-spring.datasource.url=@value.spring.datasource.url@
-spring.datasource.username=@value.spring.datasource.username@
-spring.datasource.password=@value.spring.datasource.password@
-spring.datasource.driverClassName=@value.spring.datasource.driverClassName@
-spring.jpa.database-platform=@value.spring.jpa.database-platform@
+spring.datasource.driverClassName=org.h2.Driver
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.jpa.show-sql=false
 spring.jpa.generate-ddl=true
-spring.jpa.hibernate.ddl-auto=@value.spring.jpa.hibernate.ddl-auto@
+#spring.jpa.hibernate.ddl-auto=update
