+++ passnail.security/src/main/java/com/passnail/security/service/LoginServiceIf.java
@@ -1,6 +1,5 @@
 package com.passnail.security.service;
 
-import com.passnail.data.model.entity.UserEntity;
 import com.passnail.data.transfer.model.dto.LoginDto;
 
 /**
