+++ passnail.security/src/main/test/com/passnail/security/service/impl/RegistrationServiceTest.java
@@ -1,0 +1,24 @@
+//package com.passnail.security.service.impl;
+//
+//import com.passnail.data.transfer.model.dto.RegistrationDto;
+//import com.passnail.security.SecurityModuleConfig;
+//import com.passnail.security.service.RegistrationServiceIf;
+//import org.junit.jupiter.api.Test;
+//import org.springframework.beans.factory.annotation.Autowired;
+//import org.springframework.boot.test.context.SpringBootTest;
+//
+//import static org.junit.jupiter.api.Assertions.*;
+//
+///**
+// * Created by: Pszemko at wtorek, 23.02.2021 02:34
+// * Project: passnail-client
+// */
+//@SpringBootTest(classes = {
+//        SecurityModuleConfig.class
+//})
+//public class RegistrationServiceTest {
+//
+//
+//
+//
+//}
