+++ pom.xml
@@ -35,10 +35,10 @@
         <module>passnail.data</module>
         <module>passnail.security</module>
         <module>passnail.connect</module>
-        <module>passnail.test</module>
         <module>passnail.gui</module>
         <module>passnail.generator</module>
         <module>passnail.common</module>
+        <module>passnail.test</module>
     </modules>
 
 
@@ -50,6 +50,8 @@
         <version.maven.javadoc.plugin>3.2.0</version.maven.javadoc.plugin>
         <version.dependency.check.plugin>6.1.1</version.dependency.check.plugin>
         <version.clean.plugin>3.1.0</version.clean.plugin>
+        <version.surefire.plugin>3.0.0-M5</version.surefire.plugin>
+        <version.compiler.plugin>3.8.1</version.compiler.plugin>
 
         <version.spring.boot.strater.data.jdbc>2.4.2</version.spring.boot.strater.data.jdbc>
